# Biological_app/server/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Bioserver api",
    "description": "This is Bioserver api document\n",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/grorge/bioserver/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    },
    {
      "url": "https://bioapp-backend.yikuo.dev/",
      "description": "Bioserver"
    },
    {
      "url": "http://localhost:8080",
      "description": "Bioserver local"
    }
  ],
...
  "paths": {
    "/api/login": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "登入系統",
        "operationId": "login",
        "requestBody": {
          "$ref": "#/components/requestBodies/account"
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success",
                  "data": [
                    "accessToken"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                },
                "example": [
                  {
                    "code": 1001,
                    "status": "ERROR",
                    "message": "account does not exist"
                  },
                  {
                    "code": 1002,
                    "status": "ERROR",
                    "message": "wrong password or account"
                  },
                  {
                    "code": 1004,
                    "status": "ERROR",
                    "message": " Too Many Failed Attempts (Locked after 3 failed attempts)"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/register": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "註冊系統",
        "description": "登入帳號與密碼長度規定須至少 8 碼(帳號及密碼必須限制不相同，也需要為英文數字組合)。",
        "operationId": "register",
        "requestBody": {
          "$ref": "#/components/requestBodies/account"
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                },
                "example": [
                  {
                    "code": 1005,
                    "status": "ERROR",
                    "message": "Wrong account or password format"
                  },
                  {
                    "code": 1006,
                    "status": "ERROR",
                    "message": "This account has exist"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/forgetPassword": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "忘記密碼",
        "operationId": "forgetPassword",
        "requestBody": {
          "description": "you should provide email",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_forgetPassword_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                },
                "example": [
                  {
                    "code": 1006,
                    "status": "ERROR",
                    "message": "This email has not exist"
                  },
                  {
                    "code": 400,
                    "status": "ERROR",
                    "message": "You lack necessary parameter"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/resetPassword": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "重設密碼",
        "operationId": "resetPassword",
        "requestBody": {
          "description": "you should provide email and key",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_resetPassword_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                },
                "example": [
                  {
                    "code": 1005,
                    "status": "ERROR",
                    "message": "This operation is timeout"
                  },
                  {
                    "code": 1005,
                    "status": "ERROR",
                    "message": "Wrong password format"
                  },
                  {
                    "code": 1005,
                    "status": "ERROR",
                    "message": "invalid verify"
                  }
                ]
              }
            }
          }
        }
      }
    },
    "/api/checkEmail/{code}": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "驗證email",
        "description": "將jwt的code代進params進行驗證",
        "operationId": "get_checkEmail",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful"
          }
        }
      }
    },
    "/api/checkEmail/": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "請求驗證email",
        "description": "請求驗證email,冷卻時間3分鐘",
        "operationId": "post_checkEmail",
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/api/update_password": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "修改密碼",
        "operationId": "update_password",
        "requestBody": {
          "description": "you should provide new password",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_update_password_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "example": {
                  "status": "success"
                }
              }
            }
          },
          "400": {
            "description": "wrong request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/response_ERROR"
                },
                "example": [
                  {
                    "code": 1005,
                    "status": "ERROR",
                    "message": "Wrong account or password format"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
...
    "/api/admin": {
      "get": {
        "tags": [
          "other"
        ],
        "summary": "檢查是否是管理員",
        "operationId": "get_admin",
        "parameters": [
          {
            "name": "queryname",
            "in": "query",
            "description": "admin account",
            "required": true,
            "style": "form",
            "explode": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_5"
                }
              }
            }
          },
          "400": {
            "description": "wrong request"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "other"
        ],
        "summary": "修改管理員身份",
        "operationId": "post_admin",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_admin_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "400": {
            "description": "wrong request"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "response_ERROR": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "default": "error"
          },
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "account": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
      "api_forgetPassword_body": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
      "api_resetPassword_body": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          }
        }
      },
      "api_update_password_body": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          }
        }
      },
...
      "inline_response_200_4": {
        "properties": {
          "status": {
            "type": "string",
            "default": "success"
          }
        }
      },
      "inline_response_200_5": {
        "properties": {
          "status": {
            "type": "string",
            "default": "success"
          },
          "result": {
            "type": "boolean"
          }
        }
      },
      "api_admin_body": {
        "type": "object",
        "properties": {
          "account": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          }
        }
      }
    },
    "requestBodies": {
      "account": {
        "description": "you should provide account and passowrd",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/account"
            }
          }
        },
        "required": true
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}